MVP-6451: Refactor AuthManager with some auth related type refactor
