+++ packages/notifi-frontend-client/lib/client/NotifiFrontendClient.ts
@@ -20,11 +20,12 @@ import {
 } from '../storage';
 import { notNullOrEmpty, parseTenantConfig } from '../utils';
 import { areIdsEqual } from '../utils/areIdsEqual';
-import { AuthManager, LoginParams } from './AuthManager';
+// import { AuthManager, LoginParams } from './AuthManager';
 import {
   type AlterTargetGroupParams,
   alterTargetGroupImpl,
 } from './alterTargetGroup';
+import { AuthManager, LoginParams } from './auth';
 import {
   ensureDiscord,
   ensureEmail,
