+++ packages/notifi-frontend-client/lib/client/index.ts
@@ -1,5 +1,10 @@
 export * from './NotifiFrontendClient';
 export * from './NotifiSmartLinkClient';
-export { instantiateFrontendClient, newSmartLinkClient } from './clientFactory';
+export {
+  instantiateFrontendClient,
+  newSmartLinkClient,
+  UserParams,
+} from './clientFactory';
 export * from './deprecated';
-export * from './AuthManager';
+// export * from './AuthManager';
+export * from './auth';
